Update doc string
